Cleaned up parameters with options
