--- src/main/java/ca/corefacility/bioinformatics/irida/ria/web/ajax/dto/ui/ReactComponent.java
@@ -1,7 +1,0 @@
-package ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ui;
-
-/**
- *
- */
-public class ReactComponent {
-}
