+++ src/main/webapp/resources/js/components/form/InputWithOptions.jsx
@@ -1,6 +1,16 @@
 import React from "react";
 import { Checkbox, Form, Radio, Select } from "antd";
 
+function isTruthy(options) {
+  if (options.length > 2) return false;
+  return (
+    (typeof options[0].value === "boolean" &&
+      typeof options[1].value === "boolean") ||
+    options[0].value === "true" ||
+    options[1].value === "true"
+  );
+}
+
 /**
  * React component to render an input that has pre-defined options.
  * @param item
@@ -8,16 +18,16 @@ import { Checkbox, Form, Radio, Select } from "antd";
  * @constructor
  */
 export function InputWithOptions({ item }) {
-  if (item.type === "checkbox") {
+  if (isTruthy(item.options)) {
     return (
       <Form.Item name={item.name} valuePropName="checked">
         <Checkbox>{item.label}</Checkbox>
       </Form.Item>
     );
-  } else if (item.type === "radio") {
+  } else if (item.options.length < 6) {
     return (
       <Form.Item name={item.name} label={item.label}>
-        <Radio.Group optioddns={item.options} />
+        <Radio.Group options={item.options} />
       </Form.Item>
     );
   }
