+++ src/main/webapp/resources/js/pages/launch/LaunchParametersWithOptions.jsx
@@ -16,36 +16,8 @@ import { InputWithOptions } from "../../components/form/InputWithOptions";
  * @constructor
  */
 export function LaunchParametersWithOptions({ parameters }) {
-  const [params] = React.useState(() => {
-    function isTruthy(options) {
-      if (options.length > 2) return false;
-      return (
-        (typeof options[0].value === "boolean" &&
-          typeof options[1].value === "boolean") ||
-        options[0].value === "true" ||
-        options[1].value === "true"
-      );
-    }
-    return parameters.map((p) => {
-      const parameter = { ...p };
-      if (isTruthy(parameter.options)) {
-        // Need to update to be actually boolean values
-        parameter.type = "checkbox";
-        parameter.options[0].value = parameter.options[0].value === "true";
-        parameter.options[1].value = parameter.options[0].value === "true";
-        parameter.value = parameter.value === "true";
-      } else if (parameter.options.length < 7) {
-        parameter.type = "radio";
-      } else {
-        parameter.type = "select";
-      }
-      return parameter;
-    });
-  });
-
-  const content = params.map((parameter) => (
+  const content = parameters.map((parameter) => (
     <InputWithOptions key={parameter.name} item={parameter} />
   ));
-
   return <section>{content}</section>;
 }
